Allow default access denied page to be pointed to raven/login or user/login
If we point the default access denied page to raven/login the site goes
into an infinate redirect loop.  Chaning the redirection to raven from
drupal_goto to header('Location...') fixes this.

Using $_GET['destination'] as the primary value for $redirect means
that the path that the user is trying to get to is persited.

This has been tested using raven/login and user/login as the default
path for the access denied page (site settings).  It was also been
tested with the user clicking on a link to a secure page and with a
visitor typing in an address.