Log the user out and then redirect them to a page (config) if user visits raven/logout
On Caius we want to provide a signout link to users, however as the
Raven session persists even though the user is signed out of the site
it is possible for the user to be signed back into the site without
entering their credentials again.  Hence the need to give the user a
message after they logout informing them to close their browser or
explicity log out of the site.  We could have done this with a message,
but have opted instead to redirect the user to a page.